[ADD] set and unset built-ins
