+++ src/builtins.h
@@ -31,4 +31,6 @@ int builtins_display_alias(alias_t *alias);
 int builtins_repeat(ef_t *ef, char **args);
 int builtins_yes(ef_t *ef, char **args);
 int builtins_foreach(ef_t *ef, char **args);
+int builtins_set(ef_t *ef, char **args);
+int builtins_unset(ef_t *ef, char **args);
 #endif /* BUILTIND_H */
