--- src/local/unset_local.c
@@ -1,45 +1,0 @@
-/*
-** EPITECH PROJECT, 2025
-** 42sh
-** File description:
-** unset_local
-*/
-
-#include <stdbool.h>
-#include <stddef.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "env.h"
-#include "u_mem.h"
-#include "u_str.h"
-#include "local.h"
-
-
-static
-void unset_local_move(local_t *local, size_t i)
-{
-    while (local->local_var[i]) {
-        local->local_var[i] = local->local_var[i + 1];
-        i++;
-    }
-}
-
-bool unset_local(local_t *local, char *var)
-{
-    int key_len = u_strlen(var);
-
-    for (size_t i = 0; i < local->sz; i++) {
-        if (local->local_var[i] == NULL)
-            continue;
-        if (u_strcspn(local->local_var[i], '\t') != key_len)
-            continue;
-        if (u_strcmp(local->local_var[i], var) == 0) {
-            unset_local_move(local, i);
-            local->sz--;
-            return true;
-        }
-    }
-    return false;
-}
